requirements.txt
# ==========================================
# requirements.txt
# ==========================================
fastapi
uvicorn
pillow
qrcode